requirements.txt
boto3==1.34.13
pyspark==3.5.0
requests==2.31.0
psycopg2-binary==2.9.9
awscli==1.32.13
